pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Flask-based server application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==3.0.2",
    "python-dotenv==1.0.1",
    "flask-restx"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple"
default = true
